adjust bbox size

+++ ch6_optical_flow/utils/opencv_utils.cpp
@@ -51,7 +51,7 @@ cv::Mat get_sub_image_around(cv::Mat image, int x, int y, int width, int height)
     return sub_img;
 }
 
-cv::Mat draw_bounding_box_vis_image(cv::Mat image, int x, int y, int width, int height) {
+cv::Mat draw_bounding_box_vis_image(cv::Mat image, float x, float y, float width, float height) {
     cv::rectangle(image, cv::Rect2i(x, y, width, height), cv::Scalar(0, 255, 0), 2);
     return image;
 }
