+++ ch6_optical_flow/utils/opencv_utils.h
@@ -36,7 +36,7 @@ cv::Point2i template_matching(cv::Mat img, cv::Mat temp);
  */
 cv::Mat get_sub_image_around(cv::Mat image, int x, int y, int width, int height);
 
-cv::Mat draw_bounding_box_vis_image(cv::Mat image, int x, int y, int width, int height);
+cv::Mat draw_bounding_box_vis_image(cv::Mat image, float x, float y, float width, float height);
 
 cv::Rect get_intersection_around(cv::Mat image, int x, int y, int width, int height);
